clean code

+++ console/src/main/java/com/ctrip/framework/drc/console/controller/v2/MessengerControllerV2.java
@@ -68,12 +68,15 @@ public class MessengerControllerV2 {
     @SuppressWarnings("unchecked")
     public ApiResult<Boolean> removeMessengerGroupInMha(@RequestParam String mhaName) {
         try {
+            if (defaultConsoleConfig.getNewDrcConfigSwitch().equals(DefaultConsoleConfig.SWITCH_OFF)) {
+                return ApiResult.getFailInstance(false, "not allowed");
+            }
             logger.info("removeMessengerGroupInMha in mha:{}", mhaName);
             messengerService.removeMessengerGroup(mhaName);
             return ApiResult.getSuccessInstance(true);
         } catch (Throwable e) {
             logger.error("removeMessengerGroupInMha in mha:" + mhaName, e);
-            return ApiResult.getFailInstance(null, e.getMessage());
+            return ApiResult.getFailInstance(false, e.getMessage());
         }
     }
